ChatDoc AI(mid-term project)/requirements.txt
langchain
langchain-google-genai
chainlit
faiss-cpu
sentence-transformers
PyPDF2
python-dotenv
tiktoken
duckduckgo-search

